# perf(itemsDecoding): use our own decoding algorithm for decoding items