--- a/calculators/ProfileNetworthCalculator.js
+++ b/calculators/ProfileNetworthCalculator.js
@@ -158,7 +158,6 @@ class ProfileNetworthCalculator {
                 }
 
                 // Instantiate the calculator
-                console.log(item);
                 const calculator = new calculatorClass(item);
                 // Calculate the networth of the item
                 const result = nonCosmetic