… README.md
… index.js
… types/ItemNetworthCalculator.d.ts
… types/ProfileNetworthCalculator.d.ts