Went from `550.45023ms` -> `350.8366729999998ms` on 20 averages
(`2235ms` -> `1736ms` on first run) on `56ms` 3T profile 

Also added quiver to the inventories 


```
Time taken: 2235ms
Time Taken: 519.1292319999998ms
Time Taken: 300.5708159999999ms
Time Taken: 277.90028500000017ms
Time Taken: 264.210971ms
Time Taken: 245.56890699999985ms
Time Taken: 261.82608900000014ms
Time Taken: 257.14823499999966ms
Time Taken: 241.61908200000016ms
Time Taken: 351.4515760000004ms
Time Taken: 226.9121070000001ms
Time Taken: 287.28315999999995ms
Time Taken: 221.6460449999995ms
Time Taken: 268.8136059999997ms
Time Taken: 275.3337419999998ms
Time Taken: 264.5784080000003ms
Time Taken: 249.09765599999992ms
Time Taken: 257.8096349999996ms
Time Taken: 221.69779800000015ms
Time Taken: 249.779939ms
Time Taken: 259.7711360000003ms
Average: 550.45023ms
```

```
Time taken: 1736ms
Time Taken: 396.95408599999996ms
Time Taken: 210.48277599999983ms
Time Taken: 247.95958999999993ms
Time Taken: 222.56845199999998ms
Time Taken: 180.0128239999999ms
Time Taken: 160.15028100000018ms
Time Taken: 150.77005899999995ms
Time Taken: 175.80500399999983ms
Time Taken: 236.6354650000003ms
Time Taken: 162.4349400000001ms
Time Taken: 230.61602399999992ms
Time Taken: 170.7914920000003ms
Time Taken: 155.51635699999952ms
Time Taken: 148.8438960000003ms
Time Taken: 140.5576749999991ms
Time Taken: 159.57450599999993ms
Time Taken: 157.22943300000043ms
Time Taken: 146.4480480000002ms
Time Taken: 210.50756900000033ms
Time Taken: 350.8366729999998ms
Average: 401.6895229ms
```